complete the root computation and the failure check
