+++ include/random_walks/random_walks.hpp
@@ -20,7 +20,7 @@
 #include "random_walks/uniform_john_walk.hpp"
 #include "random_walks/uniform_vaidya_walk.hpp"
 #include "random_walks/uniform_accelerated_billiard_walk.hpp"
-#include "random_walks/exponentail_hamiltonian_monte_carlo_walk.hpp"
+#include "random_walks/exponential_hamiltonian_monte_carlo_walk.hpp"
 #ifndef VOLESTIPY
     #include "random_walks/hamiltonian_monte_carlo_walk.hpp"
     #include "random_walks/langevin_walk.hpp"
