+++ include/sampling/random_point_generators.hpp
@@ -204,5 +204,78 @@ struct LogconcaveRandomPointGenerator
 };
 
 
+template
+<
+    typename Walk
+>
+struct ExponentialRandomPointGenerator
+{
+    template
+    <
+        typename Polytope,
+        typename Point,
+        typename NT,
+        typename PointList,
+        typename WalkPolicy,
+        typename RandomNumberGenerator
+    >
+    static void apply(Polytope const& P,
+                      Point &p,   // a point to start
+                      Point const& c,   // bias function
+                      NT const& T, // temperature/variance
+                      unsigned int const& rnum,
+                      unsigned int const& walk_length,
+                      PointList &randPoints,
+                      WalkPolicy &policy,
+                      RandomNumberGenerator &rng)
+    {
+        Walk walk(P, p, c, T, rng);
+        bool success;
+        for (unsigned int i=0; i<rnum; ++i)
+        {
+            success = walk.template apply(P, p, walk_length, rng);
+            if (!success) {
+                return;
+            }
+            policy.apply(randPoints, p);
+        }
+    }
+
+    template
+    <
+            typename Polytope,
+            typename Point,
+            typename NT,
+            typename PointList,
+            typename WalkPolicy,
+            typename RandomNumberGenerator,
+            typename Parameters
+    >
+    static void apply(Polytope const& P,
+                      Point &p,   // a point to start
+                      Point const& c,   // bias function
+                      NT const& T, // temperature/variance
+                      unsigned int const& rnum,
+                      unsigned int const& walk_length,
+                      PointList &randPoints,
+                      WalkPolicy &policy,
+                      RandomNumberGenerator &rng,
+                      Parameters const& parameters)
+    {
+        Walk walk(P, p, c, T, rng, parameters);
+        bool success;
+
+        for (unsigned int i=0; i<rnum; ++i)
+        {
+            success = walk.template apply(P, p, walk_length, rng);
+            if (!success) {
+                return;
+            }
+            policy.apply(randPoints, p);
+        }
+    }
+};
+
+
 
 #endif // SAMPLERS_RANDOM_POINT_GENERATORS_HPP
